requirements.txt
# ─── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.2.0
numpy>=1.26.0

# ─── LLM / Agents ─────────────────────────────────────────────────────────────
openai>=1.30.0
langgraph>=0.1.19
langchain-core>=0.2.0

# ─── Environment ──────────────────────────────────────────────────────────────
# gymnasium>=0.29.0      # optional – OpenCloudEnv is standalone, enable if needed

# ─── FAISS (DNA Memory) ───────────────────────────────────────────────────────
faiss-cpu>=1.8.0          # CPU-only; swap for faiss-gpu on hackathon VMs

# ─── UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
typing-extensions>=4.11.0

# ─── Transformers / Training (uncomment on hackathon VMs) ─────────────────────
transformers>=4.41.0
datasets>=2.19.0
# trl>=0.8.6
# unsloth                # install separately per Unsloth instructions

# ─── Experiment Tracking ──────────────────────────────────────────────────────
wandb>=0.17.0
